The game is done. I won't add save and load features, because let's be honest, who's going to save and load a game of Hearts?
The "breaking hearts before being able to start with a Hearts card" is a rule that I dislike, so I won't include it, though it should be easy to include. Other than that, the game should work perfectly fine: will test later with my friends.

Personal note if I want to come back to this project at any time in the future:
You can add save, load features
You COULD add the Hearts suit rule, depending on feedback. Just modify the playCard() method.

This was a nice learning experience, and the game turned out not bad. Some UI updates are possible too if I want in the future, though I want to focus on other stuff now as well.

Whoever visited this, thanks for exploring, coding this was kinda fun